Create bash file to automate restructuring of HTML
In many HTML files, the <meta charset = "UTF-8"> line was outside the
head, which is an incorrect file format. The script insertLine.sh takes
a line number and a final line number and a file, moving the line at
line number to the final line number and printing the result. The file
can be altered by redirecting STDOUT.